Refactor validation for hotPlugVolumes feature gate
In CNV 4.8.0, the feature gate is hardcoded by HCO to be always enabled, so there is no point to check the HCO CR for this feature gate.
Instead, we are now checking that this feature gate exists on the actual resource that makes use of it, which is the KubeVirt CR.
This is true both for CNV 2.6, and 4.8 onwards.

Signed-off-by: orenc1 <ocohen@redhat.com>