aws: remove `m4` as instanceType option on default config
The deployments in AWS region us-east-1 (biggest region) is being impacted with a cluster more expensive and without the benefits of Nitro Instances (faster CPU, network, EBS, etc). The evaluation process on that Region choose the oldest instanceType when the evaluation process of preferred zone - it expects that the instance is offered in all zones. The region us-east-1 has 6 zones, and one us-east-1e is not offering Nitro Instances. Looking the zones across all regions[1], this is the unique exception.

Regarding the performance and costs, the nitro is a better choice and it should not impact the HA on that specific zone that will use 5, instead of 6 zones.