AWS: Remove worker disk allowed types
Removing all the allowed values for workers and masters disk
types in AWS UPI. The values for disks keep changing as AWS
keeps adding types of disks fur use and it causes an
additional maintenance work for us. The disk requirements are
in the docs so users can refer to them to create the right
disk.