openstack: Inaccurate MissingQuota error
Before this patch, the OpenStack usage fetcher ("cloudinfo") conflates
the number of in-use resources with the quota and passes that to the
generic checker.

With this patch, the OpenStack usage fetcher ("cloudinfo") collects
usage and limit data in separate fields, so that the generic checker
outputs a more precise error message when the installation fails due to
insufficient resources. For example: "the required number of resources
(1499) is more than remaining quota of 1498" instead of "the required
number of resources (1499) is more than the limit of 1498".