Bug 2010665: do not modify cvo ignores for bootstrap-in-place
When doing a bootstrap-in-place installation, the cluster-version-operator
does not start until after the bootstrap machine has pivoted to being a
control-plane machine. Consequently, there is no possibility of the
cluster-bootstrap racing with the cluster-version-operator to create
resources. The code that was added to fix Bug 2008119 [1] is not necessary
for bootstrap-in-place installations. In fact, the code actually breaks
bootstrap-in-place installations by trying to restore the CVO overrides
after cluster-bootstrap finishes but before the bootstrap-to-control-plane
pivot occurs.

The changes in this commit restrict the code for Bug 2008119 to only
installations that are not bootstrap-in-place.

[1] https://github.com/openshift/installer/pull/5258