aws: Remove shared tags from user specified IAM roles
The installer tags the IAM roles that are specified by the user
and deletes it during cluster destroy which is not ideal since
the installer should not be deleting the user created resources.

Removing the code that tags the IAM roes provided by the user and
also removing the permission required to delete the IAM role if
the user provided it.