pkg/asset/manifests/operators: Set kubernetes.io/description for cluster-config-v1
The ConfigMap is install-time archival, and may no longer express the
current cluster state.  Having it in-cluster was intended to be a
temporary hack until we got Infrastructure and such, and we just failed
to remove the ConfigMap quickly enough, and now removing it would be a
breaking change.  But with this commit, we get a warning about this
into the cluster, so folks looking at the resource are more likely to
understand its limitations.

The kubernetes.io/description annotation has been documented since [1].

The Annotations property brings over the relevant parts of:

  $ git --no-pager grep 'Annotations.*map.*json' vendor/k8s.io/apimachinery/pkg/apis/meta/v1
  vendor/k8s.io/apimachinery/pkg/apis/meta/v1/types.go:   Annotations map[string]string `json:"annotations,omitempty" protobuf:"bytes,12,rep,name=annotations"`

[1]: https://github.com/kubernetes/website/pull/31621